+++ AmazonChimeSDK/AmazonChimeSDK/internal/audio/DefaultAudioClientObserver.swift
@@ -269,12 +269,13 @@ class DefaultAudioClientObserver: NSObject, AudioClientDelegate {
             return
         }
 
-        DispatchQueue.global().async {
-            self.audioLock.lock()
-            self.audioClient.stopSession()
+        DispatchQueue.global().async { [weak self] in
+            guard let strongSelf = self else { return }
+            strongSelf.audioLock.lock()
+            strongSelf.audioClient.stopSession()
             DefaultAudioClientController.state = .stopped
-            self.audioLock.unlock()
-            self.notifyAudioClientObserver { (observer: AudioVideoObserver) in
+            strongSelf.audioLock.unlock()
+            strongSelf.notifyAudioClientObserver { (observer: AudioVideoObserver) in
                 observer.audioSessionDidStopWithStatus(sessionStatus: MeetingSessionStatus(statusCode: newAudioStatus))
             }
         }